requirements.txt
sleekxmpp >= 1.3.1
python-telegram-bot >= 4.0.3
